:white_check_mark: test: 구매금액에 따른 랜덤 로또 생성 및 출력 로직 테스트 통과
