+++ src/main/java/lotto/controller/Game.java
@@ -5,12 +5,13 @@ import lotto.domain.Lotto;
 import lotto.parser.Parser;
 import lotto.util.Generator;
 import lotto.view.InputView;
+import lotto.view.OutputView;
 
 public class Game {
     public static void start() {
         String purchaseAmount = InputView.requestPurchaseAmount();
         int ticketAmount = Parser.parsePurchaseAmount(purchaseAmount);
-        List<Lotto> ticket = Generator.publishTicket(ticketAmount);
-
+        List<Lotto> tickets = Generator.publishTickets(ticketAmount);
+        OutputView.responseTickets(ticketAmount, tickets);
     }
 }
