+++ src/main/java/lotto/view/OutputView.java
@@ -5,9 +5,9 @@ import static lotto.view.constants.ConstantMessage.RESPONSE_FIRST;
 import static lotto.view.constants.ConstantMessage.RESPONSE_FOURTH;
 import static lotto.view.constants.ConstantMessage.RESPONSE_HYPHEN_SEPARATOR;
 import static lotto.view.constants.ConstantMessage.RESPONSE_PROFIT_RATE;
-import static lotto.view.constants.ConstantMessage.RESPONSE_PURCHASE_AMOUNT;
 import static lotto.view.constants.ConstantMessage.RESPONSE_SECOND;
 import static lotto.view.constants.ConstantMessage.RESPONSE_THIRD;
+import static lotto.view.constants.ConstantMessage.RESPONSE_TICKET_AMOUNT;
 import static lotto.view.constants.ConstantMessage.RESPONSE_WINNING_STATISTICS;
 
 import java.util.List;
@@ -15,10 +15,10 @@ import lotto.domain.Lotto;
 import lotto.view.constants.ConstantMessage;
 
 public class OutputView {
-    public static void responsePurchaseAmount(int amount, List<Lotto> automaticTickets) {
-        ConstantMessage.printConstantMessage(RESPONSE_PURCHASE_AMOUNT, amount);
-        for (Lotto automaticTicket : automaticTickets) {
-            System.out.println(automaticTicket.getNumbers());
+    public static void responseTickets(int ticketAmount, List<Lotto> tickets) {
+        ConstantMessage.printConstantMessage(RESPONSE_TICKET_AMOUNT, ticketAmount);
+        for (Lotto ticket : tickets) {
+            System.out.println(ticket.getNumbers());
         }
     }
