+++ src/main/java/lotto/view/constants/ConstantMessage.java
@@ -4,7 +4,7 @@ public enum ConstantMessage {
     REQUEST_PURCHASE_AMOUNT("구입금액을 입력해 주세요."),
     REQUEST_WINNING_NUMBERS("당첨 번호를 입력해 주세요."),
     REQUEST_BONUS_NUMBER("보너스 번호를 입력해 주세요."),
-    RESPONSE_PURCHASE_AMOUNT("%d개를 구매했습니다."),
+    RESPONSE_TICKET_AMOUNT("%d개를 구매했습니다."),
     RESPONSE_WINNING_STATISTICS("당첨 통계"),
     RESPONSE_HYPHEN_SEPARATOR("---"),
     MATCHING_WINNING_NUMBER("개 일치"),
